LPS-150150 Move PasswordEncryptor impls to module
