+++ modules/apps/portal-security/portal-security-password-encryptor-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay Portal Security Password Encryptor Implementation
+Bundle-SymbolicName: com.liferay.portal.security.password.encryptor.impl
+Bundle-Version: 1.0.0
