+++ modules/apps/portal-security/portal-security-password-encryptor-impl/src/main/java/com/liferay/portal/security/password/encryptor/internal/CompositePasswordEncryptor.java
@@ -12,8 +12,10 @@
  * details.
  */
 
-package com.liferay.portal.security.pwd;
+package com.liferay.portal.security.password.encryptor.internal;
 
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
+import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
@@ -26,13 +28,15 @@ import com.liferay.portal.kernel.util.ClassUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.util.PropsValues;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import org.osgi.framework.BundleContext;
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 
 /**
  * @author Michael C. Han
  */
+@Component(property = "composite=true", service = PasswordEncryptor.class)
 public class CompositePasswordEncryptor
 	extends BasePasswordEncryptor implements PasswordEncryptor {
 
@@ -127,17 +131,15 @@ public class CompositePasswordEncryptor
 		throw new UnsupportedOperationException();
 	}
 
-	public void setPasswordEncryptors(
-		List<PasswordEncryptor> passwordEncryptors) {
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		_passwordEncryptors = ServiceTrackerMapFactory.openSingleValueMap(
+			bundleContext, PasswordEncryptor.class, "type");
+	}
 
-		for (PasswordEncryptor passwordEncryptor : passwordEncryptors) {
-			if (_log.isDebugEnabled()) {
-				_log.debug("Registering " + passwordEncryptor);
-			}
-
-			_passwordEncryptors.put(
-				passwordEncryptor.getAlgorithmType(), passwordEncryptor);
-		}
+	@Deactivate
+	protected void deactivate() {
+		_passwordEncryptors.close();
 	}
 
 	private String _getAlgorithmName(String algorithm) {
@@ -158,15 +160,15 @@ public class CompositePasswordEncryptor
 		PasswordEncryptor passwordEncryptor = null;
 
 		if (algorithm.startsWith(PasswordEncryptorUtil.TYPE_BCRYPT)) {
-			passwordEncryptor = _passwordEncryptors.get(
+			passwordEncryptor = _passwordEncryptors.getService(
 				PasswordEncryptorUtil.TYPE_BCRYPT);
 		}
 		else if (algorithm.startsWith(PasswordEncryptorUtil.TYPE_PBKDF2)) {
-			passwordEncryptor = _passwordEncryptors.get(
+			passwordEncryptor = _passwordEncryptors.getService(
 				PasswordEncryptorUtil.TYPE_PBKDF2);
 		}
 		else {
-			passwordEncryptor = _passwordEncryptors.get(algorithm);
+			passwordEncryptor = _passwordEncryptors.getService(algorithm);
 		}
 
 		if (passwordEncryptor == null) {
@@ -174,7 +176,7 @@ public class CompositePasswordEncryptor
 				_log.debug("No password encryptor found for " + algorithm);
 			}
 
-			passwordEncryptor = _passwordEncryptors.get(
+			passwordEncryptor = _passwordEncryptors.getService(
 				PasswordEncryptorUtil.TYPE_DEFAULT);
 		}
 
@@ -191,7 +193,6 @@ public class CompositePasswordEncryptor
 	private static final Log _log = LogFactoryUtil.getLog(
 		CompositePasswordEncryptor.class);
 
-	private final Map<String, PasswordEncryptor> _passwordEncryptors =
-		new HashMap<>();
+	private ServiceTrackerMap<String, PasswordEncryptor> _passwordEncryptors;
 
 }
