+++ modules/apps/portal-security/portal-security-password-encryptor-impl/src/main/java/com/liferay/portal/security/password/encryptor/internal/DefaultPasswordEncryptor.java
@@ -12,57 +12,35 @@
  * details.
  */
 
-package com.liferay.portal.security.pwd;
+package com.liferay.portal.security.password.encryptor.internal;
 
 import com.liferay.portal.kernel.security.pwd.PasswordEncryptor;
 import com.liferay.portal.kernel.security.pwd.PasswordEncryptorUtil;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.kernel.util.DigesterUtil;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import jodd.util.BCrypt;
+import org.osgi.service.component.annotations.Component;
 
 /**
  * @author Michael C. Han
  * @author Tomas Polesovsky
  */
-public class BCryptPasswordEncryptor
+@Component(
+	property = "type=" + PasswordEncryptorUtil.TYPE_DEFAULT,
+	service = PasswordEncryptor.class
+)
+public class DefaultPasswordEncryptor
 	extends BasePasswordEncryptor implements PasswordEncryptor {
 
 	@Override
 	public String encrypt(
 		String algorithm, String plainTextPassword, String encryptedPassword) {
 
-		String salt = null;
-
-		if (Validator.isNull(encryptedPassword)) {
-			int rounds = _ROUNDS;
-
-			Matcher matcher = _pattern.matcher(algorithm);
-
-			if (matcher.matches()) {
-				rounds = GetterUtil.getInteger(matcher.group(1), rounds);
-			}
-
-			salt = BCrypt.gensalt(rounds);
-		}
-		else {
-			salt = encryptedPassword.substring(0, 29);
-		}
-
-		return BCrypt.hashpw(plainTextPassword, salt);
+		return DigesterUtil.digest(algorithm, plainTextPassword);
 	}
 
 	@Override
 	public String getAlgorithmType() {
-		return PasswordEncryptorUtil.TYPE_BCRYPT;
+		return PasswordEncryptorUtil.TYPE_DEFAULT;
 	}
-
-	private static final int _ROUNDS = 10;
-
-	private static final Pattern _pattern = Pattern.compile(
-		"^BCrypt/([0-9]+)$", Pattern.CASE_INSENSITIVE);
 
 }
